chore(eslint-config): update configurations and rules following dependency upgrade
Updated ESLint configurations and rules in response to the latest version upgrade of ESLint and associated plugins. This change includes adjustments to property naming conventions to align with new API expectations and removal of deprecated options. The refactor ensures compatibility with the updated dependencies and maintains the existing code standards enforced by our ESLint rules.