chore(nuxt): update `ContextMenu` component styling and interaction
Updated the `ContextMenu` component in the Nuxt examples to enhance its usability and visual appeal. Changed the position of the menu to appear below and to the left of the trigger area, improving accessibility and intuitiveness. Additionally, replaced the previous transition implementation with `BaseCollapse` for smoother animations and added conditional classes to manage the visual state based on whether the menu is open. This ensures a more consistent and visually appealing interaction for users.