fix(vue): adjust initial state setting to prevent unwanted transitions on mount
Updated the `BaseCollapse` component to prevent unwanted transitions when the component is first mounted. Removed the call to `setStyleContent` in the `onMounted` lifecycle hook and refined the initial state setting of the `style` ref. This ensures that the transition effects only apply when there is an actual state change after the component has been initialized, avoiding any unintended animations during the initial render.