chore(client): streamline imports in `createClient.ts`
Adjusted the import statements in `createClient.ts` to improve code clarity and maintainability. Moved the importing of `OpenAPIV2` and `OpenAPIV3` types to the top with other type imports and simplified the import from `./openapi/index` by removing type imports that were redundantly specified in multiple lines. This change makes the import section cleaner and more organized.